pyproject.toml
[tool.poetry]
name = "lang-chain-learning"
version = "0.1.0"
description = ""
authors = ["Nathan Stuart <nathan.stuart01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <3.13"
tiktoken = "^0.6.0"
pandas = "^2.2.0"
openai = "^1.12.0"
bs4 = "^0.0.2"
pinecone-client = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
